refactor: give the hvPlot views their own module
base.py served every view type and had grown to 2000 lines, of which the
hvPlot views and their constants were a quarter. They move to
lumen/views/hvplot.py together, so the styling options added here sit
with the views that use them rather than among the deck.gl and Vega-Lite
ones.

The module needs an __all__: it imports hvplot itself, and the package's
star import would otherwise rebind that name over the submodule.
