+++ lumen/tests/ai/test_agents.py
@@ -586,7 +586,7 @@ async def test_analysis_agent(llm, duckdb_source, test_messages):
     class TestAnalysis(Analysis):
 
         def __call__(self, pipeline, context):
-            return f"Test Analysis"
+            return "Test Analysis"
 
     agent = AnalysisAgent(
         analyses=[TestAnalysis.instance(name='foo'), TestAnalysis.instance(name='bar')],
@@ -630,17 +630,17 @@ class TestDocumentListAgentIntegration:
                 DocumentChunk(filename="schema.md", text="chunk 3", similarity=0.7),
             ]
         )
-        
+
         context = {"metaset": metaset}
-        
+
         # Test applies
         applies = await DocumentListAgent.applies(context)
         assert applies is True  # More than 1 unique document
-        
+
         # Test _get_items
         agent = DocumentListAgent()
         items = agent._get_items(context)
-        
+
         # Should return unique, sorted filenames
         assert items == {"Documents": ["readme.md", "schema.md"]}
 
@@ -649,7 +649,7 @@ class TestDocumentListAgentIntegration:
         # Metaset without docs
         metaset = Metaset(query="test", catalog={}, docs=None)
         context = {"metaset": metaset}
-        
+
         applies = await DocumentListAgent.applies(context)
         assert applies is False
